refactor: MoveCount 객체 생성
