+++ src/main/java/racingcar/domain/MoveCount.java
@@ -1,9 +1,16 @@
-package racingcar.validator;
+package racingcar.domain;
 
-public class MoveCountValidator {
+public class MoveCount {
 	public static final String ERROR_NOT_POSITIVE = "[ERROR] 이동횟수는 1 이상의 정수여야 합니다.";
 
-	public static void validateMoveCount(int moveCount) {
+	private final int moveCount;
+
+	public MoveCount(int moveCount) {
+		validateMoveCount(moveCount);
+		this.moveCount = moveCount;
+	}
+
+	private static void validateMoveCount(int moveCount) {
 		checkPositiveInteger(moveCount);
 	}
 
@@ -11,5 +18,9 @@ public class MoveCountValidator {
 		if (moveCount <= 0) {
 			throw new IllegalArgumentException(ERROR_NOT_POSITIVE);
 		}
+	}
+
+	public int toInt() {
+		return moveCount;
 	}
 }
